% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmConstraint.R
\name{adjacent-ergmHint}
\alias{adjacent-ergmHint}
\alias{InitErgmConstraint.adjacent}
\alias{adjacent-ergmConstraint}
\title{Try to make jumps as small as possible}
\usage{
# adjacent
}
\description{
Statistic calculation is more efficient if the proposed value of an edge is as close as possible to its original value.
}
\seealso{
\code{\link[ergm:ergmHint]{ergmHint}} for index of constraints and hints currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmHint", "adjacent", "subsection")}
}
\concept{dyad-independent}
