% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{AdjacentAlterSwap-ergmProposal}
\alias{AdjacentAlterSwap-ergmProposal}
\alias{InitWtErgmProposal.AdjacentAlterSwap}
\title{A proposal that swaps values of two adjacently-ranked alters incident on an ego}
\description{
This proposal randomly selects a dyad \eqn{(i,j)}, then finds an alter \eqn{j'} ranked immediately above \eqn{j}, and proposes to swap their values.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="AdjacentAlterSwap"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="AdjacentAlterSwap"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="AdjacentAlterSwap"))}}
}
\seealso{
\code{\link[ergm:ergmProposal]{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "AdjacentAlterSwap", "subsection")}
}
\concept{ordinal}
\concept{valued}
\keyword{internal}
