% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{AlterSwap-ergmProposal}
\alias{AlterSwap-ergmProposal}
\alias{InitWtErgmProposal.AlterSwap}
\title{A proposal that swaps values of two alters incident on an ego}
\description{
This proposal randomly selects two dyads \eqn{(i,j)} and \eqn{(i,j')} with a common sender and proposes to swap their values if distinct.
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="AlterSwap"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="AlterSwap"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="AlterSwap"))}}
}
\seealso{
\code{\link{ergmProposal}} for index of proposals currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmProposal", "AlterSwap", "subsection")}
}
\concept{ordinal}
\concept{valued}
\keyword{internal}
