% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmProposal.R
\name{AlterSwap-ergmProposal}
\alias{AlterSwap-ergmProposal}
\alias{InitWtErgmProposal.AlterSwap}
\title{TODO}
\description{
TODO
}
\details{
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsHtml(ergm:::.buildProposalsList(proposal="AlterSwap"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsText(ergm:::.buildProposalsList(proposal="AlterSwap"))}}
\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatProposalsLatex(ergm:::.buildProposalsList(proposal="AlterSwap"))}}
}
