% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitErgmReference.R
\name{CompleteOrder-ergmReference}
\alias{CompleteOrder-ergmReference}
\alias{InitErgmReference.CompleteOrder}
\title{A uniform distribution over the possible complete orderings of the alters by each ego}
\usage{
# CompleteOrder
}
\description{
A uniform distribution over the possible complete orderings of the alters by each ego
}
\seealso{
\code{\link{ergmReference}} for index of reference distributions currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmReference", "CompleteOrder", "subsection")}
}
\concept{ordinal}
\concept{valued}
