% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.rank-terms-index.R
\docType{package}
\name{ergmReference}
\alias{ergmReference}
\alias{ergm-references}
\alias{references-ergm}
\alias{ergm.references}
\alias{references.ergm}
\alias{InitWtErgmProposal.AlterSwap}
\title{Reference Measures for Exponential-Family Random Graph Models for Rank-Order
Relational Data}
\description{
This page describes the possible reference measures (baseline distributions)
for modeling rank data found in the \code{\link[=ergm.rank-package]{ergm.rank}} package.

Each of these is specified on the RHS of a one-sided formula passed as the
\code{reference} argument to \code{\link{ergm}}. See the \code{\link{ergm}} documentation for a complete description of how
reference measures are specified.
}
\section{Possible reference measures to represent baseline distributions}{


Reference measures currently available are:
\itemize{
\item \code{DiscUnif(a,b)}: A discrete uniform distribution used as a
baseline distribution for ranks. See \code{\link[=DiscUnif-ergmReference]{DiscUnif}} documentation in the
\code{\link[ergm:ergm-package]{ergm}} for arguments. Using
\code{reference=~DiscUnif(1,n-1)} (for network size \code{n}) can be used to
model weak orderings, though this approach is currently untested. Note that
it entails a specific assumption about actors' propensity to rank.
}
\subsection{Reference index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmReference"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmReference"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmReference"))}}
}
}

\references{
\itemize{
\item Krivitsky PN (2012). Exponential-Family Random Graph Models for
Valued Networks. \emph{Electronic Journal of Statistics}, 2012, 6,
1100-1128. \doi{10.1214/12-EJS696}
\item Krivitsky PN and Butts CT (2017). Exponential-Family Random Graph Models for
Rank-Order Relational Data. \emph{Sociological Methodology}, 2017, 47,
68-112. \doi{10.1177/0081175017692623}
}
}
\seealso{
\code{\link[=ergm-package]{ergm}}, \code{\link{network}}, \verb{\\\%v\\\%}, \verb{\\\%n\\\%}, \code{sna}, \code{\link{summary.ergm}}, \code{\link{print.ergm}}
}
\keyword{models}
