% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ergm.rank-terms-index.R
\docType{package}
\name{ergmTerm}
\alias{ergmTerm}
\alias{ergm-terms}
\alias{ergm.terms}
\alias{terms-ergm}
\alias{terms.ergm}
\title{Terms used in Exponential Family Random Graph Models for Ranks}
\description{
This page describes the possible terms (and hence network statistics)
included in the \code{\link[=ergm.rank-package]{ergm.rank}} package.

See the \code{\link[=ergmTerm]{ergm}} documentation in the \code{ergm}
package for a complete description of what ERGM terms are and how they are
used.
}
\section{Terms to represent network statistics included in the \code{\link[=ergm.rank-package]{ergm.rank}} pacakge}{

All terms listed are valued. For their specific forms, see Krivitsky and
Butts (2012).

These terms have a specialized interpretation, and are therefore generally
prefixed by "\code{rank.}", though they should take any valued data.
\subsection{Term index}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexLatex(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexText(ergm:::.buildTermsDataframe("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatIndexHtml(ergm:::.buildTermsDataframe("ergmTerm"))}}
}

\subsection{Frequently-used terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::FREQUENTLY_USED_TERM_CATEGORIES, only.include='frequently-used'))}}
}

\subsection{Operator terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm", categories=ergm:::OPERATOR_CATEGORIES, only.include='operator'))}}
}

\subsection{All terms}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixLatex(ergm:::.termMatrix("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixText(ergm:::.termMatrix("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatMatrixHtml(ergm:::.termMatrix("ergmTerm"))}}
}

\subsection{Terms by concept}{

\if{latex}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocLatex(ergm:::.termToc("ergmTerm"))}}
\if{text}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocText(ergm:::.termToc("ergmTerm"))}}
\if{html}{\Sexpr[results=rd,stage=render]{ergm:::.formatTocHtml(ergm:::.termToc("ergmTerm"))}}
}
}

\references{
\itemize{
\item Krivitsky P. N. (2012). Exponential-Family Random Graph Models for
Valued Networks. \emph{Electronic Journal of Statistics}, 6, 1100-1128.
\doi{10.1214/12-EJS696}
\item Krivitsky PN and Butts CT (2017). Exponential-Family Random Graph
Models for Rank-Order Relational Data. \emph{Sociological Methodology},
2017, 47, 68-112. \doi{10.1177/0081175017692623}
}
}
\seealso{
\code{\link[ergm:ergm-terms]{ergm-terms}} (from the
\code{\link[=ergm-package]{ergm}} package), \code{\link{ergm}}, \code{\link{network}}, \verb{\%v\%}, \verb{\%n\%}
}
\keyword{models}
