% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm_rank.R
\name{rank.nodeicov-ergmTerm}
\alias{rank.nodeicov-ergmTerm}
\alias{InitWtErgmTerm.rank.nodeicov}
\title{Attractiveness/Popularity covariates}
\usage{
# valued: rank.nodeicov(attr)
}
\arguments{
\item{attr}{quantitative attribute (see Specifying Vertex attributes and Levels (\code{?nodal_attributes}) for details.)}
}
\description{
Models the effect of one or more nodal covariates on the propensity of an
actor to be ranked highly by the others.
}
\concept{directed}
\concept{ordinal}
\concept{quantitative nodal attribute}
\concept{valued}
