% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm_rank.R
\name{rank.deference-ergmTerm}
\alias{rank.deference-ergmTerm}
\alias{InitWtErgmTerm.rank.deference}
\title{Deference (aversion)}
\usage{
# valued: rank.deference
}
\description{
Measures the
amount of "deference" in the network: configurations where an ego
\eqn{i} ranks an alter \eqn{j} over another alter \eqn{k}, but
\eqn{j}, in turn, ranks \eqn{k} over \eqn{i} . A lower-than-chance
value of this statistic and/or a negative coefficient implies a form
of mutuality in the network.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "rank.deference", "subsection")}
}
\concept{directed}
\concept{ordinal}
\concept{triad-related}
\concept{valued}
