% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitWtErgmTerm_rank.R
\name{rank.edgecov-ergmTerm}
\alias{rank.edgecov-ergmTerm}
\alias{InitWtErgmTerm.rank.edgecov}
\title{Dyadic covariates}
\usage{
# valued: rank.edgecov(x, attrname)
}
\arguments{
\item{x, attrname}{either a square matrix of covariates, one for
each possible edge in the network, the name of a network
attribute of covariates, or a network; if the latter, or if the
network attribute named by \code{x} is itself a network, optional
argument \code{attrname} provides the name of the quantitative edge
attribute to use for covariate values (in this case, missing
edges in \code{x} are assigned a covariate value of zero).}
}
\description{
Models the effect of a dyadic covariate on the propensity of an ego
\eqn{i} to rank alter \eqn{j} highly.
}
\seealso{
\code{\link{ergmTerm}} for index of model terms currently visible to the package.

\Sexpr[results=rd,stage=render]{ergm:::.formatTermKeywords("ergmTerm", "rank.edgecov", "subsection")}
}
\concept{directed}
\concept{ordinal}
\concept{quantitative dyadic attribute}
\concept{valued}
